필드 명ì이나 파일 명 ã파일 ì위치 등을 DDD 형식에 맞게 수정
