+++ BookManagement/src/main/java/com/example/demo/book/repository/BookRepository.java
@@ -1,0 +1,13 @@
+package com.example.demo.book.repository;
+
+import com.example.demo.book.entity.BookEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface BookRepository extends JpaRepository<BookEntity, Long> {
+    Optional<BookEntity> findByBookId(Long bookId);
+
+}
