--- BookManagement/src/main/java/com/example/demo/entity/RentalEntity.java
@@ -1,34 +1,0 @@
-package com.example.demo.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Setter
-@Table(name = "rental_table")
-public class RentalEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long rentalId;
-
-    @Column
-    private LocalDateTime rentedAt;
-
-    @Column
-    private LocalDateTime dueAt;
-
-    @Column
-    private LocalDateTime returnedAt;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private UserEntity user;
-
-    @ManyToOne
-    @JoinColumn(name = "book_id")
-    private BookEntity book;
-}
