--- BookManagement/src/main/java/com/example/demo/repository/BookRepository.java
@@ -1,13 +1,0 @@
-package com.example.demo.repository;
-
-import com.example.demo.entity.BookEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface BookRepository extends JpaRepository<BookEntity, Long> {
-    Optional<BookEntity> findByBookId(Long bookId);
-
-}
