--- BookManagement/src/main/java/com/example/demo/repository/RentalRepository.java
@@ -1,15 +1,0 @@
-package com.example.demo.repository;
-
-import com.example.demo.entity.RentalEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface RentalRepository extends JpaRepository<RentalEntity, Long> {
-    Optional<RentalEntity>  findByRentalId(Long rentalId);
-
-    List<RentalEntity> findByUser_UserIdAndReturnedAtIsNull(Long userId);
-}
