--- BookManagement/src/main/java/com/example/demo/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.example.demo.repository;
-
-import com.example.demo.entity.MemberEntity;
-import com.example.demo.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-    Optional<UserEntity> findByUserEmail(String userEmail);
-}
