+++ BookManagement/src/main/java/com/example/demo/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.example.demo.user.repository;
+
+import com.example.demo.user.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+    Optional<UserEntity> findByUserEmail(String userEmail);
+//    Optional<UserEntity> findByUserName(String userName);
+}
