--- BookManagement/src/main/resources/templates/List.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <title>Title</title>
-</head>
-<body>
-<h2>list.html</h2>
-<table>
-  <tr>
-    <th>id</th>
-    <th>memberEmail</th>
-    <th>memberPassword</th>
-    <th>memberName</th>
-    <th>상세조회</th>
-    <th>삭제</th>
-  </tr>
- <!-- th:each로 묶음을 만든다. -->
-  <tr th:each="member: ${memberDTOList}">
-    <td th:text="${member.id}"></td>
-    <td th:text="${member.memberEmail}"></td>
-    <td th:text="${member.memberPassword}"></td>
-    <td th:text="${member.memberName}"></td>
-    <td>
-      <!-- query string /member?id=1
-          rest api /member/1 -->
-      <!--타임리프 문법의 |를 통해 서버에서 가져온 값을 같이 표현 가능하다.-->
-      <a th:href="@{|/member/${member.id}|}">조회</a>
-    </td>
-    <td>
-      <a th:href="@{|/member/delete/${member.id}|}">삭제</a>
-    </td>
-  </tr>
-</table>
-</body>
-</html>
