+++ BookManagement/src/main/resources/templates/admin/adminBookList.html
@@ -1,0 +1,118 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+  <meta charset="UTF-8">
+  <title>📚 관리자 책 목록</title>
+  <style>
+    body {
+      font-family: 'Noto Sans KR', sans-serif;
+      background-color: #f8f9fa;
+      margin: 0;
+      padding: 0;
+    }
+
+    h2 {
+      text-align: center;
+      margin-top: 30px;
+      font-size: 22px;
+      color: #333;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      gap: 20px;
+    }
+
+    .delete-btn {
+      padding: 6px 14px;
+      background-color: crimson;
+      color: white;
+      border: none;
+      border-radius: 4px;
+      cursor: pointer;
+      font-size: 14px;
+    }
+
+    .delete-btn:hover {
+      background-color: darkred;
+    }
+
+    table {
+      width: 90%;
+      margin: 20px auto;
+      border-collapse: collapse;
+      font-size: 14px;
+      background-color: white;
+    }
+
+    th, td {
+      padding: 12px;
+      border: 1px solid #ddd;
+      text-align: center;
+    }
+
+    th {
+      background-color: #f4f4f4;
+      font-size: 15px;
+    }
+
+    tr:hover {
+      background-color: #f0f8ff;
+    }
+  </style>
+
+  <script>
+    function submitDelete() {
+      const checkboxes = document.querySelectorAll('input[name="bookIds"]:checked');
+      if (checkboxes.length === 0) {
+        alert("삭제할 책을 한 개 이상 선택해주세요.");
+        return;
+      }
+      document.getElementById("deleteForm").submit();
+    }
+
+    function rowClick(bookId) {
+      window.location.href = "/book/update/" + bookId;
+    }
+
+    function toggleAll(source) {
+      const checkboxes = document.querySelectorAll('input[name="bookIds"]');
+      checkboxes.forEach(cb => cb.checked = source.checked);
+    }
+  </script>
+</head>
+<body>
+
+<form id="deleteForm" th:action="@{/admin/book}" method="post">
+  <h2>
+    <button type="button" class="delete-btn" onclick="submitDelete()">선택 삭제</button>
+    📚 책 목록
+  </h2>
+
+  <table>
+    <thead>
+    <tr>
+      <th><input type="checkbox" onclick="toggleAll(this)"></th>
+      <th>제목</th>
+      <th>국제번호</th>
+      <th>대여 가능 재고/전체 재고</th>
+      <th>출판사</th>
+      <th>저자</th>
+    </tr>
+    </thead>
+    <tbody>
+    <tr th:each="book : ${books}" th:onclick="'rowClick(' + ${book.bookId} + ')'" style="cursor: pointer;">
+      <td>
+        <input type="checkbox" name="bookIds" th:value="${book.bookId}" onclick="event.stopPropagation()">
+      </td>
+      <td th:text="${book.bookTitle}"></td>
+      <td th:text="${book.bookIsbn}"></td>
+      <td th:text="${book.bookCurrentStock + ' / ' + book.bookOriginalStock}"></td>
+      <td th:text="${book.bookPublishedAt}"></td>
+      <td th:text="${book.bookPublisher}"></td>
+    </tr>
+    </tbody>
+  </table>
+</form>
+
+</body>
+</html>
