--- BookManagement/src/main/resources/templates/adminUserList.html
@@ -1,118 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>👤 관리자 유저 목록</title>
-    <style>
-        body {
-          font-family: 'Noto Sans KR', sans-serif;
-          background-color: #f8f9fa;
-          margin: 0;
-          padding: 0;
-        }
-
-        h2 {
-          text-align: center;
-          margin-top: 30px;
-          font-size: 22px;
-          color: #333;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          gap: 20px;
-        }
-
-        .delete-btn {
-          padding: 6px 14px;
-          background-color: crimson;
-          color: white;
-          border: none;
-          border-radius: 4px;
-          cursor: pointer;
-          font-size: 14px;
-        }
-
-        .delete-btn:hover {
-          background-color: darkred;
-        }
-
-        table {
-          width: 90%;
-          margin: 20px auto;
-          border-collapse: collapse;
-          font-size: 14px;
-          background-color: white;
-        }
-
-        th, td {
-          padding: 12px;
-          border: 1px solid #ddd;
-          text-align: center;
-        }
-
-        th {
-          background-color: #f4f4f4;
-          font-size: 15px;
-        }
-
-        tr:hover {
-          background-color: #f0f8ff;
-        }
-    </style>
-
-    <script>
-        function submitDelete() {
-          const checkboxes = document.querySelectorAll('input[name="userIds"]:checked');
-          if (checkboxes.length === 0) {
-            alert("삭제할 사용자를 한 명 이상 선택해주세요.");
-            return;
-          }
-          document.getElementById("deleteForm").submit();
-        }
-
-        function rowClick(userId) {
-          window.location.href = "/user/update/" + userId;
-        }
-
-        function toggleAll(source) {
-          const checkboxes = document.querySelectorAll('input[name="userIds"]');
-          checkboxes.forEach(cb => cb.checked = source.checked);
-        }
-    </script>
-</head>
-<body>
-
-<form id="deleteForm" th:action="@{/admin/user}" method="post">
-    <h2>
-        <button type="button" class="delete-btn" onclick="submitDelete()">선택 삭제</button>
-        👤 유저 목록
-    </h2>
-
-    <table>
-        <thead>
-        <tr>
-            <th><input type="checkbox" onclick="toggleAll(this)"></th>
-            <th>이메일</th>
-            <th>닉네임</th>
-            <th>전화번호</th>
-            <th>가입일</th>
-            <th>관리자 여부</th>
-        </tr>
-        </thead>
-        <tbody>
-        <tr th:each="user : ${users}">
-            <td>
-                <input type="checkbox" name="userIds" th:value="${user.userId}" onclick="event.stopPropagation()">
-            </td>
-            <td th:text="${user.userEmail}"></td>
-            <td th:text="${user.userName}"></td>
-            <td th:text="${user.userPhoneNum}"></td>
-            <td th:text="${#temporals.format(user.userCreated, 'yyyy-MM-dd HH:mm')}"></td>
-            <td th:text="${user.admin ? '관리자' : '일반'}"></td>
-        </tr>
-        </tbody>
-    </table>
-</form>
-
-</body>
-</html>
