--- BookManagement/src/main/resources/templates/bookSave.html
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko" xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <title>책 등록</title>
-  <style>
-    body {
-        text-align: center;
-        margin-top: 50px;
-        font-family: Arial, sans-serif;
-    }
-    h1 {
-        margin-bottom: 30px;
-    }
-    form {
-        display: inline-block;
-        text-align: left;
-        padding: 20px;
-        border: 1px solid #ccc;
-        border-radius: 8px;
-        background-color: #f9f9f9;
-    }
-    .form-group {
-        margin-bottom: 15px;
-    }
-    label {
-        display: inline-block;
-        width: 100px;
-    }
-    input[type="text"],
-    input[type="number"],
-    select {
-        width: 250px;
-        padding: 5px;
-        font-size: 14px;
-    }
-    .btn-submit {
-        margin-top: 20px;
-        width: 100%;
-        padding: 10px;
-        font-size: 16px;
-        background-color: #4CAF50;
-        color: white;
-        border: none;
-        border-radius: 5px;
-        cursor: pointer;
-    }
-    .btn-submit:hover {
-        background-color: #45a049;
-    }
-  </style>
-</head>
-
-<body>
-
-<h1>책 등록</h1>
-
-<form action="/book/save" method="post">
-  <div class="form-group">
-    <label for="bookTitle">제목:</label>
-    <input type="text" id="bookTitle" name="bookTitle" required>
-  </div>
-  <div class="form-group">
-    <label for="bookIsbn">국제번호:</label>
-    <input type="text" id="bookIsbn" name="bookIsbn" required>
-  </div>
-
-  <div class="form-group">
-    <label for="bookOriginalStock">재고:</label>
-    <input type="number" id="bookOriginalStock" name="bookOriginalStock" required>
-
-  </div>
-  <div class="form-group">
-    <label for="bookPublisher">발행자:</label>
-    <input type="text" id="bookPublisher" name="bookPublisher" required>
-  </div>
-  <div class="form-group">
-    <label for="bookPublished_at">출판사:</label>
-    <select id="bookPublished_at" name="bookPublished_at" required>
-      <option value=" ">선택하세요</option>
-      <option value="한아름">한아름</option>
-      <option value="두마리">두마리</option>
-      <option value="삼손">삼손</option>
-    </select>
-  </div>
-  <button type="submit" class="btn-submit">책 등록</button>
-</form>
-
-</body>
-</html>
