--- BookManagement/src/main/resources/templates/bookUpdate.html
@@ -1,106 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <title>📘 책 정보 수정</title>
-  <style>
-    body {
-      font-family: 'Noto Sans KR', sans-serif;
-      background-color: #f4f4f4;
-      margin: 0;
-      padding: 0;
-    }
-
-    .container {
-      width: 400px;
-      margin: 50px auto;
-      padding: 20px;
-      background-color: white;
-      border-radius: 10px;
-      box-shadow: 0 2px 8px rgba(0,0,0,0.1);
-    }
-
-    h2 {
-      text-align: center;
-      color: #333;
-      margin-bottom: 20px;
-    }
-
-    label {
-      display: block;
-      margin-top: 10px;
-      font-weight: bold;
-    }
-
-    input, select {
-      width: 100%;
-      padding: 8px;
-      margin-top: 5px;
-      border: 1px solid #ccc;
-      border-radius: 4px;
-    }
-
-    .btn-submit {
-      width: 100%;
-      margin-top: 20px;
-      padding: 10px;
-      background-color: #007bff;
-      color: white;
-      border: none;
-      border-radius: 4px;
-      font-size: 16px;
-      cursor: pointer;
-    }
-
-    .btn-submit:hover {
-      background-color: #0056b3;
-    }
-
-    a {
-      display: block;
-      text-align: center;
-      margin-top: 15px;
-      color: #007bff;
-      text-decoration: none;
-    }
-
-    a:hover {
-      text-decoration: underline;
-    }
-  </style>
-</head>
-<body>
-<div class="container">
-  <h2>책 정보 수정</h2>
-  <form th:action="@{/book/update}" method="post">
-    <input type="hidden" name="bookId" th:value="${book.bookId}" />
-
-    <label>제목</label>
-    <input type="text" name="bookTitle" th:value="${book.bookTitle}" required/>
-
-    <label>국제번호</label>
-    <input type="text" name="bookIsbn" th:value="${book.bookIsbn}" required/>
-
-    <label>전체 재고</label>
-    <input type="text" name="bookOriginalStock" th:value="${book.bookOriginalStock}" required/>
-
-    <label>현재 재고</label>
-    <input type="text" name="bookCurrentStock" th:value="${book.bookCurrentStock}" required/>
-
-    <label>발행자</label>
-    <input type="text" name="bookPublished_at" th:value="${book.bookPublished_at}" required/>
-
-    <label>출판사</label>
-    <select name="bookPublisher" required>
-      <option th:selected="${book.bookPublisher == '한아름'}" value="한아름">한아름</option>
-      <option th:selected="${book.bookPublisher == '두마리'}" value="두마리">두마리</option>
-      <option th:selected="${book.bookPublisher == '삼손'}" value="삼손">삼손</option>
-    </select>
-
-    <button type="submit" class="btn-submit">수정 완료</button>
-  </form>
-
-  <a href="/admin/book">목록으로 돌아가기</a>
-</div>
-</body>
-</html>
