--- BookManagement/src/main/resources/templates/detail.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <title>detail</title>
-</head>
-<body>
-<table>
-  <tr>
-    <th>id</th>
-    <th>memberEmail</th>
-    <th>memberPassword</th>
-    <th>memberName</th>
-  </tr>
-  <tr>
-    <td th:text="${member.id}"></td>
-    <td th:text="${member.memberEmail}"></td>
-    <td th:text="${member.memberPassword}"></td>
-    <td th:text="${member.memberName}"></td>
-  </tr>
-</table>
-</body>
-</html>
