+++ BookManagement/src/main/resources/templates/user/userSave.html
@@ -1,0 +1,87 @@
+<!DOCTYPE html>
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>회원가입</title>
+    <style>
+        body {
+            text-align: center;
+            margin-top: 50px;
+            font-family: Arial, sans-serif;
+        }
+        h1 {
+            margin-bottom: 30px;
+        }
+        form {
+            display: inline-block;
+            text-align: left;
+            padding: 20px;
+            border: 1px solid #ccc;
+            border-radius: 8px;
+            background-color: #f9f9f9;
+        }
+        .form-group {
+            margin-bottom: 15px;
+        }
+        label {
+            display: inline-block;
+            width: 100px;
+        }
+        input[type="text"],
+        input[type="email"],
+        input[type="password"],
+        input[type="tel"] {
+            width: 250px;
+            padding: 5px;
+            font-size: 14px;
+        }
+        .btn-submit {
+            margin-top: 20px;
+            width: 100%;
+            padding: 10px;
+            font-size: 16px;
+            background-color: #4CAF50;
+            color: white;
+            border: none;
+            border-radius: 5px;
+            cursor: pointer;
+        }
+        .btn-submit:hover {
+            background-color: #45a049;
+        }
+    </style>
+</head>
+
+<body>
+
+<h1>회원가입</h1>
+
+<form th:action="@{${actionUrl}}" method="post" th:object="${user}">
+
+    <div class="form-group">
+        <label for="userEmail">이메일:</label>
+        <input type="email" id="userEmail" name="userEmail" required>
+    </div>
+
+    <div class="form-group">
+        <label for="userPassword">Password:</label>
+        <input type="password" id="userPassword" name="userPassword" required>
+    </div>
+
+    <div class="form-group">
+        <label for="userName">이름:</label>
+        <input type="text" id="userName" name="userName">
+    </div>
+    <div class="form-group">
+        <label for="userPhoneNumber">전화번호:</label>
+        <input type="tel" id="userPhoneNumber" name="userPhoneNumber">
+    </div>
+    <button type="submit"
+            th:text="${isAdmin} ? '관리자 가입' : '회원 가입'"
+            th:style="${isAdmin} ? 'background-color: blue; color: white;' : ''">
+        회원 가입
+    </button>
+</form>
+
+</body>
+</html>
