--- BookManagement/src/main/resources/templates/userRentals.html
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <title>내 대여 목록</title>
-  <style>
-    table {
-      width: 80%;
-      margin: 30px auto;
-      border-collapse: collapse;
-    }
-    th, td {
-      padding: 10px;
-      border: 1px solid #ddd;
-      text-align: center;
-    }
-    th {
-      background-color: #f2f2f2;
-    }
-    .return-btn {
-      padding: 5px 10px;
-      background-color: #2196F3;
-      color: white;
-      border: none;
-      border-radius: 4px;
-      cursor: pointer;
-    }
-    .return-btn:hover {
-      background-color: #0b7dda;
-    }
-  </style>
-</head>
-<body>
-
-<h2 style="text-align:center">📖 내 대여 목록</h2>
-
-<div th:if="${message}" style="text-align:center; color: red;" th:text="${message}"></div>
-
-<table>
-  <thead>
-  <tr>
-    <th>책 제목</th>
-    <th>출판사</th>
-    <th>대여일</th>
-    <th>반납 예정일</th>
-    <th>반납</th>
-  </tr>
-  </thead>
-  <tbody>
-  <tr th:each="rental : ${rentals}">
-    <td th:text="${rental.book.bookTitle}">제목</td>
-    <td th:text="${rental.book.bookPublisher}">출판사</td>
-    <td th:text="${#temporals.format(rental.rentedAt, 'yyyy-MM-dd')}">대여일</td>
-    <td th:text="${#temporals.format(rental.dueAt, 'yyyy-MM-dd')}">반납예정</td>
-    <td>
-      <form th:action="@{'/book/return/' + ${rental.rentalId}}" method="post">
-        <button type="submit" class="return-btn">반납하기</button>
-      </form>
-    </td>
-  </tr>
-  </tbody>
-</table>
-
-</body>
-</html>
